requirements.txt
requests
beautifulsoup4
markdownify
python-dotenv
langchain-core
langchain-google-genai
pydantic
django
django-cors-headers